+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/depcies/ch_view_deps_change_table_usr_t4_new.sql
@@ -1,0 +1,8 @@
+CREATE TABLE default.t4
+(
+	`Id` UInt32 NOT NULL,
+	`Object.Value` Array(String) NOT NULL
+)
+ENGINE = MergeTree
+ORDER BY Id
+SETTINGS index_granularity = 8192;