+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/parsers/antlr/chexpr/ChAbstractExprWithNmspc.java
@@ -23,14 +23,20 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
-import java.util.stream.Stream;
 
-import ru.taximaxim.codekeeper.core.PgDiffUtils;
+import org.antlr.v4.runtime.ParserRuleContext;
+
+import ru.taximaxim.codekeeper.core.Consts;
+import ru.taximaxim.codekeeper.core.model.difftree.DbObjType;
+import ru.taximaxim.codekeeper.core.parsers.antlr.QNameParser;
+import ru.taximaxim.codekeeper.core.parsers.antlr.generated.CHParser.Alias_clauseContext;
 import ru.taximaxim.codekeeper.core.parsers.antlr.generated.CHParser.Dml_stmtContext;
+import ru.taximaxim.codekeeper.core.parsers.antlr.generated.CHParser.Qualified_nameContext;
 import ru.taximaxim.codekeeper.core.parsers.antlr.generated.CHParser.With_clauseContext;
 import ru.taximaxim.codekeeper.core.parsers.antlr.generated.CHParser.With_queryContext;
 import ru.taximaxim.codekeeper.core.schema.GenericColumn;
 import ru.taximaxim.codekeeper.core.schema.IRelation;
+import ru.taximaxim.codekeeper.core.schema.PgObjLocation.LocationType;
 import ru.taximaxim.codekeeper.core.schema.meta.MetaContainer;
 import ru.taximaxim.codekeeper.core.utils.Pair;
 
@@ -54,12 +60,20 @@ public abstract class ChAbstractExprWithNmspc<T> extends ChAbstractExpr {
      */
     private final Set<String> cte = new HashSet<>();
 
+    /**
+     * This variable is used to isolate references at current level.
+     * We need it to not lose the dependencies in with_clauseContext
+     */
+    private final boolean isLocalScope;
+
     protected ChAbstractExprWithNmspc(String schema, MetaContainer meta) {
         super(schema, meta);
+        this.isLocalScope = false;
     }
 
-    protected ChAbstractExprWithNmspc(ChAbstractExpr parent) {
+    protected ChAbstractExprWithNmspc(ChAbstractExpr parent, boolean isLocalScope) {
         super(parent);
+        this.isLocalScope = isLocalScope;
     }
 
     @Override
@@ -73,7 +87,7 @@ public abstract class ChAbstractExprWithNmspc<T> extends ChAbstractExpr {
         return ref == null ? super.findReferenceRecursive(schema, name) : ref;
     }
 
-    protected Entry<String, GenericColumn> findReferenceInNmspc(String schema, String name) {
+    private Entry<String, GenericColumn> findReferenceInNmspc(String schema, String name) {
         boolean found;
         GenericColumn dereferenced = null;
         if (schema == null && namespace.containsKey(name)) {
@@ -104,6 +118,31 @@ public abstract class ChAbstractExprWithNmspc<T> extends ChAbstractExpr {
         return found ? new SimpleEntry<>(name, dereferenced) : null;
     }
 
+    @Override
+    protected void addReferenceInRootParent(Qualified_nameContext name, Alias_clauseContext alias, boolean isFrom) {
+        if (isLocalScope || !hasParent()) {
+            addNameReference(name, alias, isFrom);
+        } else {
+            super.addReferenceInRootParent(name, alias, isFrom);
+        }
+    }
+
+    /**
+     * Clients may use this to setup pseudo-variable names before expression
+     * analysis.
+     */
+    private void addReference(String alias, GenericColumn object) {
+        if (namespace.containsKey(alias)) {
+            log(Consts.DUPLICATE_ALIASES, alias);
+            return;
+        }
+        namespace.put(alias, object);
+    }
+
+    public void addReference(String alias) {
+        addReference(alias, null);
+    }
+
     public boolean addRawTableReference(GenericColumn qualifiedTable) {
         boolean exists = !unaliasedNamespace.add(qualifiedTable);
         if (exists) {
@@ -123,18 +162,18 @@ public abstract class ChAbstractExprWithNmspc<T> extends ChAbstractExpr {
 
     private Pair<IRelation, Pair<String, String>> findColumn(String name, Collection<GenericColumn> refs) {
         for (GenericColumn ref : refs) {
-            if (ref == null) {
-                continue;
-            }
-            IRelation rel = findRelation(ref.schema, ref.table);
-            if (rel == null) {
-                continue;
-            }
+            if (ref != null) {
+                IRelation rel = findRelation(ref.schema, ref.table);
+                if (rel == null) {
+                    continue;
+                }
 
-            Stream<Pair<String, String>> columns = rel.getRelationColumns();
-            for (Pair<String, String> col : PgDiffUtils.sIter(columns)) {
-                if (col.getFirst().equals(name)) {
-                    return new Pair<>(rel, col);
+                Pair<IRelation, Pair<String, String>> pair = rel.getRelationColumns()
+                        .filter(e -> e.getFirst().equals(name))
+                        .findFirst().map(e -> new Pair<>(rel, e))
+                        .orElse(null);
+                if (pair != null) {
+                    return pair;
                 }
             }
         }
@@ -142,6 +181,10 @@ public abstract class ChAbstractExprWithNmspc<T> extends ChAbstractExpr {
     }
 
     protected void analyzeCte(With_clauseContext with) {
+        if (with == null) {
+            return;
+        }
+
         for (With_queryContext withQuery : with.with_query()) {
             String withName = withQuery.name.getText();
             var expr = withQuery.expr();
@@ -150,7 +193,6 @@ public abstract class ChAbstractExprWithNmspc<T> extends ChAbstractExpr {
             } else {
                 Dml_stmtContext data = withQuery.dml_stmt();
                 var select = data.select_stmt();
-                // TODO add other later
                 if (select != null) {
                     new ChSelect(this).analyze(select);
                 }
@@ -160,6 +202,38 @@ public abstract class ChAbstractExprWithNmspc<T> extends ChAbstractExpr {
                 log("Duplicate CTE " + withName);
             }
         }
+    }
+
+    private GenericColumn addNameReference(Qualified_nameContext name, Alias_clauseContext alias, boolean isFrom) {
+        String firstName = QNameParser.getFirstName(name.identifier());
+        boolean isCte = name.DOT().isEmpty() && hasCte(firstName);
+        GenericColumn depcy = null;
+        if (!isCte && isFrom) {
+            depcy = addObjectDepcy(name, DbObjType.TABLE);
+        }
+
+        if (alias != null) {
+            ParserRuleContext aliasCtx = getAliasCtx(alias);
+            if (depcy != null) {
+                // add alias definition
+                addReference(depcy, aliasCtx, LocationType.VARIABLE);
+            }
+            addReference(aliasCtx.getText(), depcy);
+        } else if (isCte) {
+            addReference(firstName, null);
+        } else {
+            addRawTableReference(depcy);
+        }
+
+        return depcy;
+    }
+
+    protected ParserRuleContext getAliasCtx(Alias_clauseContext alias) {
+        ParserRuleContext aliasCtx = alias.identifier();
+        if (aliasCtx == null) {
+            aliasCtx = alias.id_token();
+        }
+        return aliasCtx;
     }
 
     public abstract List<String> analyze(T ruleCtx);